Added method to easier allow additional post types to be added
Zoninator #12

Added a `register_zone_post_type` method to the `Zoninator` class. It more easily allows users to add additional post types to zones.

It accepts one parameter, `$post_types`, and it can be either a string of the post type slug, or an array of post type slugs.

The method returns a boolean: true if any post types were registered, false if not.