Add post meta box
Added the meta box and the save method to all admin post pages that can be used in Zones. Only some user roles can assign posts to Zones, and this is filtered. Users who cannot will see the list of Zones still as read only. Also, only published and future post statuses can be added to Zones. Moving a post to a non-published state will remove it from all Zones.